git-svn-id: http://svn.lee.org/swarm/trunk@285 47a5675e-1c2e-0410-904b-cc838fe3ce0b

+++ spu/include/swarmdefines.h
@@ -103,7 +103,6 @@ struct	swarmGpsData
 
 struct	swarmImuData
 {
-
   /* IMU data converted to SI units */
   double si_ratex;
   double si_ratey;
@@ -120,12 +119,12 @@ struct	swarmImuData
 
   /* integer bias (what we get when we measure a zero reading */
   /* subtract these values from the raw int vals to get a zero-mean value*/
-  int ratex_bias;
-  int ratey_bias;
-  int accx_bias;
-  int accy_bias;
-  int accz_bias;
-
+  /* [even though these are doubles, their units are still raw (NOT SI!)] */
+  double ratex_bias;
+  double ratey_bias;
+  double accx_bias;
+  double accy_bias;
+  double accz_bias;
 };
 
 struct  swarmStateEstimate
